Collections refactor
The TypeCollection is removed in favor of using a type to identify the collection items.

The collections got a more consistent language.
So is there a createWith[X] methods for the collections that often require the creation and calling the add method in the same operation.
The add method is to create a new item instance with the instance arguments.
To created instance needs to be added to the collection, the add[X] method is available.

The Alias class is added to make it possible to filter the data value if it needs it.

Several functions are changed to use the changed collection methods and new functionality.